+++ src/components/molecules/TimerSideBar/index.tsx
@@ -23,7 +23,7 @@ const TimerSideBar = ({ ongoingTodos = todoData, completedTodos = todoData, togg
 	const { animate, handleClose } = useCloseSidebar(toggleSidebar);
 	return (
 		<div
-			className={`t flex h-screen w-[40.2rem] transform flex-col rounded-bl-[16px] rounded-tl-[16px] bg-gray-bg-03 pl-[1.8rem] transition-transform duration-300 ease-in-out ${animate ? 'translate-x-0' : 'translate-x-full'}`}
+			className={`flex h-[108rem] w-[40.2rem] transform flex-col rounded-bl-[16px] rounded-tl-[16px] bg-gray-bg-03 pl-[1.8rem] transition-transform duration-300 ${animate ? 'translate-x-0' : 'translate-x-full'}`}
 		>
 			<div className="flex h-[5.4rem] w-[36.6rem] items-center justify-between pl-[0.2rem] pt-[2rem]">
 				<p className="head-bold-24 text-white">오늘 할 일</p>
