+++ src/components/molecules/TimerTitle/index.tsx
@@ -1,18 +1,5 @@
-interface TodoTitleProps {
-	title: string;
-}
-
-const TodoTitle = ({ title }: TodoTitleProps) => {
-	return <h1 className="title-semibold-64 text-white">{title}</h1>;
-};
-
-interface CategoryTitleProps {
-	title: string;
-}
-
-const CategoryTitle = ({ title }: CategoryTitleProps) => {
-	return <h2 className="title-med-32 text-gray-04">{title}</h2>;
-};
+import CategoryTitle from './CategoryTitle';
+import TodoTitle from './TodoTitle';
 
 const TimerTitle = () => {
 	return (
