+++ src/pages/TimerPage/index.tsx
@@ -14,24 +14,26 @@ const TimerPage = () => {
 
 	return (
 		<TimerPageTemplates>
-			<div>
-				<div className="flex h-[108rem] w-[192rem] bg-[url('@/assets/images/img_timer_bg.png')]">
+			<div className="relative flex h-[108rem] w-[192rem] bg-[url('@/assets/images/img_timer_bg.png')]">
+				<div className="absolute left-0">
 					<TimerSideBox />
-					<div className="ml-[54.5rem] mt-[-0.8rem]">
-						<TimerTitle />
-						<Timer />
-						<FriendInfoCarousel />
-					</div>
-					<button
-						onClick={toggleSidebar}
-						className="absolute right-0 top-0 mr-[3.4rem] mt-[3.2rem] rounded-[1.5rem] hover:bg-gray-bg-04"
-					>
-						<HamburgerIcon />
-					</button>
 				</div>
+				<div className="ml-[56.6rem] mt-[-0.8rem]">
+					<TimerTitle />
+					<Timer />
+					<FriendInfoCarousel />
+				</div>
+				<button
+					onClick={toggleSidebar}
+					className="ml-[38.2rem] mt-[3.2rem] h-[5.4rem] w-[5.4rem] rounded-[1.5rem] hover:bg-gray-bg-04"
+				>
+					<HamburgerIcon />
+				</button>
 				{isSidebarOpen && (
-					<div className={`fixed inset-0 flex justify-end bg-gray-bg-01 bg-opacity-50`}>
-						<TimerSideBar toggleSidebar={toggleSidebar} />
+					<div className="absolute inset-0 z-10 bg-dim">
+						<div className="absolute inset-y-0 right-0 flex justify-end overflow-hidden">
+							<TimerSideBar toggleSidebar={toggleSidebar} />
+						</div>
 					</div>
 				)}
 			</div>
